requirements.txt
nose
oyaml
pandas
pathos
numpy >= 1.13.0
root-numpy >= 4.6.0
#pyhf
